+++ src/main/java/com/garnerju/auctionhouse/repository/BidRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BidRepository extends JpaRepository<Bid, Double> {
+public interface BidRepository extends JpaRepository<Bid, Long> {
 }
