+++ src/main/java/com/garnerju/auctionhouse/repository/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Double> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
